fix: read the Simple preset values of the Secondary role from default user configuration
It also applys Max [suggestion](https://github.com/UltimateHackingKeyboard/agent/issues/2909#issuecomment-4459830980) to keep custom preset unchanged until the user does not modify the user configuration
