+++ packages/uhk-common/src/index.ts
@@ -1,4 +1,5 @@
 export { Buffer } from './buffer.js';
+export * from './config-serializer/config-items/advanced-secondary-role-configuration-values.js';
 export * from './config-serializer/index.js';
 export * from './log/index.js';
 export * from './models/index.js';
