+++ packages/uhk-common/user-config-80.json
@@ -65,12 +65,12 @@
   "mouseScrollAcceleratedSpeed": 50,
   "secondaryRoleStrategy": "Simple",
   "secondaryRoleAdvancedStrategyDoubletapTimeout": 200,
-  "secondaryRoleAdvancedStrategyTimeout": 350,
-  "secondaryRoleAdvancedStrategySafetyMargin": 50,
-  "secondaryRoleAdvancedStrategyTrigger": "Release",
-  "secondaryRoleAdvancedStrategyDoubletapToPrimary": true,
+  "secondaryRoleAdvancedStrategyTimeout": 1000,
+  "secondaryRoleAdvancedStrategySafetyMargin": 0,
+  "secondaryRoleAdvancedStrategyTrigger": "Press",
+  "secondaryRoleAdvancedStrategyDoubletapToPrimary": false,
   "secondaryRoleAdvancedStrategyTimeoutAction": "Secondary",
-  "secondaryRoleAdvancedStrategyTriggerByMouse": false,
+  "secondaryRoleAdvancedStrategyTriggerByMouse": true,
   "secondaryRoleAdvancedStrategyTriggerFromSameHalf": true,
   "secondaryRoleAdvancedStrategyMinimumHoldTime": 0,
   "secondaryRoleAdvancedStrategyTimeoutType": "Active",
